+++ src/handlers/user.rs
@@ -39,7 +39,7 @@ pub async fn list_user_transactions(Path(id): Path<Uuid>) -> impl IntoResponse {
       id: Some(Uuid::new_v4()),
       sender: Some(id),
       receiver: Some(Uuid::new_v4()),
-      amount: 1000,
+      amount: 1000 as f64,
       currency: Some(Currencies::USD),
       // This hash example is wrong, just for test
       hash: Uuid::new_v4().to_string(),
